qdrant-cookbook/requirements.txt
honeyhive>=1.0.0rc21
openai
qdrant-client
openinference-instrumentation-openai
python-dotenv
